Move schema base modules to __init__.py.
