+++ mash/services/api/schema/accounts/__init__.py
@@ -1,0 +1,29 @@
+# Copyright (c) 2019 SUSE LLC.  All rights reserved.
+#
+# This file is part of mash.
+#
+# mash is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# mash is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with mash.  If not, see <http://www.gnu.org/licenses/>
+#
+
+from mash.services.api.schema import non_empty_string
+
+delete_account = {
+    'type': 'object',
+    'properties': {
+        'account_name': non_empty_string,
+        'requesting_user': non_empty_string,
+    },
+    'additionalProperties': False,
+    'required': ['account_name', 'requesting_user']
+}
