+++ mash/services/api/schema/accounts/ec2.py
@@ -18,7 +18,7 @@
 
 import copy
 
-from mash.services.api.schema.base import non_empty_string
+from mash.services.api.schema import non_empty_string
 
 additional_regions = {
     'type': 'array',
