+++ mash/services/api/schema/accounts/gce.py
@@ -18,7 +18,7 @@
 
 import copy
 
-from mash.services.api.schema.base import non_empty_string
+from mash.services.api.schema import non_empty_string
 
 gce_account = {
     'type': 'object',
