+++ mash/services/api/schema/jobs/ec2.py
@@ -18,8 +18,8 @@
 
 import copy
 
-from mash.services.api.schema.base import non_empty_string
-from mash.services.api.schema.jobs.base import base_job_message
+from mash.services.api.schema import non_empty_string
+from mash.services.api.schema.jobs import base_job_message
 
 account_numbers = {
     'type': 'string',
